Move XHR hook logic to xhrHook.js
